+++ src/Firebase/ServiceAccount.php
@@ -37,6 +37,11 @@ class ServiceAccount
         return $serviceAccount;
     }
 
+    public function hasClientId(): bool
+    {
+        return (bool) $this->clientId;
+    }
+
     public function getClientId(): string
     {
         return $this->clientId;
@@ -64,6 +69,11 @@ class ServiceAccount
         $serviceAccount->clientEmail = $value;
 
         return $serviceAccount;
+    }
+
+    public function hasPrivateKey(): bool
+    {
+        return (bool) $this->privateKey;
     }
 
     public function getPrivateKey(): string
@@ -136,6 +146,15 @@ class ServiceAccount
         } catch (\Throwable $e) {
             throw new InvalidArgumentException(sprintf('%s can not be read.', $filePath));
         }
+    }
+
+    public static function withProjectIdAndServiceAccountId(string $projectId, string $serviceAccountId): self
+    {
+        $serviceAccount = new self();
+        $serviceAccount->projectId = $projectId;
+        $serviceAccount->clientEmail = $serviceAccountId;
+
+        return $serviceAccount;
     }
 
     /**